Add NetCDF-C++ port and beginnings of dTect plugin
Changes to use the latest version of NetCDF-C++ which
is better designed, more "C++-like" and can make use
of the latest versions of the C library.

Some documentation is provided for this.

Also adds an initial plugin framework for OpendTect,
in line for future integration between Jif3D and dTect.